Correct the package suffix for all configs
The package will not build in "Release" because the library suffix ("Project description" option under "Project options") is set to "D10" and in "Debug" is set to "D110". Thus, the compiler will throw an error that can't find the file "*_D110.bpl". Set the library suffix to "D110" for "All platforms - All configurations" and remove this from the "Debug" configuration. Now, the package compiles in both "Release" and "Debug" configurations.