Added ScaleMems function with Pascal and SSE4.1 implementations.
ScaleMems can be used to fade an image to black. See the ParticleSwarm and GradLines examples.
For fading, ScaleMems is preferred over BlendMems since it's impossible to fade to complete black with BlendMems due to rounding errors. ScaleMems is also slightly faster (~5%).
For now ScaleMems is not guaranteed to preserve the alpha channel.
Added benchmark mode to ParticleSwarm example.