Implemented region blur replacements.
The region blur is a generic solution implemented with a bitmap pattern polygon filler.
Additionally a bug has been fixed in the gamma variant of the old region motion blur; It didn't constrain the blur to the specified region which indicates that this function has never been used by anyone.