1) Fixed alpha handling bug in kernel resamplers.
When the source bitmap did not have alpha in the range 1..254 the resampler assumed that it didn't need to premultiply alpha. However, this didn't take into account that a source bitmap with both alpha=0 and alpha=255 would produce target values with alpha in the 1..254 range and would thus need premultiplication.
2) Added alpha-aware version of interpolator.